% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlur.R
\name{blurContainer}
\alias{blurContainer}
\title{Blur container}
\usage{
blurContainer(tag)
}
\arguments{
\item{tag}{Element to apply the blur effect.}
}
\description{
Allow to apply a blur effect on a given element.
}
