% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPersp.R
\name{setPersp}
\alias{setPersp}
\title{Custom perspective effect}
\usage{
setPersp(
  id = NULL,
  class = NULL,
  direction = "Y",
  angle = 45,
  depth = 600,
  side = "left",
  hover = FALSE
)
}
\arguments{
\item{id}{Use this argument if you want to target an individual element.}

\item{class}{The element to which the persp should be applied.}

\item{direction}{Perspective direction: "X" or "Y".}

\item{angle}{Rotation intensity: numeric, 45 by default.}

\item{depth}{Perspective depth: numeric, 600 by default.}

\item{side}{Perspective side: "left" or "right".}

\item{hover}{Whether to apply the persp effect on hover. FALSE by default.}
}
\description{
Allow to apply a perspective effect on a given element.
}
\examples{
 if (interactive()) {

  library(shiny)
  library(shinydashboard)
  library(shinydashboardPlus)
  library(shinyEffects)

  boxTag <- boxPlus(
   title = "Closable box, with label",
   closable = TRUE,
   enable_label = TRUE,
   label_text = 1,
   label_status = "danger",
   status = "warning",
   solidHeader = FALSE,
   collapsible = TRUE,
   p("Box Content")
  )

  shinyApp(
   ui = dashboardPagePlus(
     header = dashboardHeaderPlus(
       enable_rightsidebar = TRUE,
       rightSidebarIcon = "gears"
     ),
     sidebar = dashboardSidebar(),
     body = dashboardBody(

      setPersp(id = "my-box-1", angle = 45),
      setPersp(id = "my-box-2", angle = 45, side = "right", hover = TRUE),

      tags$h2("Add Perspective effects"),
      br(),
      fluidRow(
       tagAppendAttributes(boxTag[[2]], id = "my-box-1"),
       tagAppendAttributes(boxTag[[2]], id = "my-box-2")
      )
     ),
     rightsidebar = rightSidebar(),
     title = "DashboardPage"
   ),
   server = function(input, output) { }
  )
}
}
