% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEffectsGallery.R
\name{shinyEffectsGallery}
\alias{shinyEffectsGallery}
\title{Launch the shinyEffects Gallery}
\usage{
shinyEffectsGallery()
}
\description{
A gallery of all components available in shinyEffects.
}
\examples{

if (interactive()) {

 shinyEffectsGallery()

}
}
