% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{use_effects}
\alias{use_effects}
\title{shinyEffects}
\usage{
use_effects()
}
\description{
Use only if you need shinyEffects on the server side
}
\examples{
library(shiny)

ui <- fluidPage(
  use_effects(), # dependencies

  actionButton("show", "Show loading for 5 seconds")
)

server <- function(input, output, session){
  waiter_hide() # will hide *on_load waiter

  observeEvent(input$show, {

  })
}

if(interactive()) shinyApp(ui, server)

}
